media-source/SourceBuffer-appendStreamWithMaxSize_updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#appendStream(stream, maxSize) when updating attribute is true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'stream';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(xhr.response);
    };
}
//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET(FILE, function(stream) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updatestart', function() {
                t.step(function() {
                    assert_throws({name: 'InvalidStateError'},
                                  function() {
                                      sourceBuffer.appendStream(stream, 1024);
                                  });
                });
                t.done();
            });
            //add dummy buffer to make updating attribute true
            sourceBuffer.appendBuffer(new Uint8Array(8));
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, 'SourceBuffer#appendStream(stream, maxSize) : If the updating attribute equals true, INVALID_STATE_ERR exception should be thrown.');

</script>
</body>
</html>
